pyproject.toml
[project]
name = "skinwalker"
version = "0.1.0"
description = "Terminal skin studio for Hermes CLI skins"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "Pillow>=11.0.0",
  "PyYAML>=6.0.2",
  "pyfiglet>=1.0.2",
  "rich>=13.7.1",
  "textual>=0.38.0",
]

[project.scripts]
skinwalker = "skinwalker.__main__:main"
imagewalker = "imagewalker.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/skinwalker", "src/imagewalker"]

[tool.uv]
package = true
